refactor: Float 컴포넌트
