+++ src/hooks/useDirection.ts
@@ -1,0 +1,39 @@
+import { DIRECTION } from "@/pages/mypage/review";
+import { useEffect, useState } from "react";
+
+const useDirection = () => {
+  const [scrollDir, setScrollDir] = useState<DIRECTION>();
+
+  useEffect(() => {
+    const threshold = 0;
+    let lastScrollY = window.pageYOffset;
+    let ticking = false;
+
+    const updateScrollDir = () => {
+      const scrollY = window.pageYOffset;
+
+      if (Math.abs(scrollY - lastScrollY) < threshold) {
+        ticking = false;
+        return;
+      }
+      setScrollDir(scrollY > lastScrollY ? DIRECTION.Down : DIRECTION.Up);
+      lastScrollY = scrollY > 0 ? scrollY : 0;
+      ticking = false;
+    };
+
+    const onScroll = () => {
+      if (!ticking) {
+        window.requestAnimationFrame(updateScrollDir);
+        ticking = true;
+      }
+    };
+
+    window.addEventListener("scroll", onScroll);
+
+    return () => window.removeEventListener("scroll", onScroll);
+  }, [scrollDir]);
+
+  return [scrollDir];
+};
+
+export default useDirection;
