+++ src/pages/mypage/review/index.tsx
@@ -3,48 +3,20 @@ import Layout from "@/components/Layout";
 import HeaderDepth from "@/components/Layout/Header/HeaderDepth";
 import styled from "@emotion/styled";
 import React, { useEffect, useState } from "react";
-import Button from "@/components/Button";
 import Link from "next/link";
 import { Post, posts } from "@/mock/posts";
 import Icon from "@/util/Icon";
+import Float from "@/components/UI/Float";
+import useDirection from "@/hooks/useDirection";
 
-const DIRECTION = {
+export const DIRECTION = {
   Up: "UP",
   Down: "Down",
 } as const;
-type DIRECTION = typeof DIRECTION[keyof typeof DIRECTION]; // 'UP' | DOWN'
+export type DIRECTION = typeof DIRECTION[keyof typeof DIRECTION]; // 'UP' | DOWN'
 
 const Review = () => {
-  const [scrollDir, setScrollDir] = useState<DIRECTION>();
-
-  useEffect(() => {
-    const threshold = 0;
-    let lastScrollY = window.pageYOffset;
-    let ticking = false;
-
-    const updateScrollDir = () => {
-      const scrollY = window.pageYOffset;
-
-      if (Math.abs(scrollY - lastScrollY) < threshold) {
-        ticking = false;
-        return;
-      }
-      setScrollDir(scrollY > lastScrollY ? DIRECTION.Down : DIRECTION.Up);
-      lastScrollY = scrollY > 0 ? scrollY : 0;
-      ticking = false;
-    };
-
-    const onScroll = () => {
-      if (!ticking) {
-        window.requestAnimationFrame(updateScrollDir);
-        ticking = true;
-      }
-    };
-
-    window.addEventListener("scroll", onScroll);
-
-    return () => window.removeEventListener("scroll", onScroll);
-  }, [scrollDir]);
+  const [scrollDir] = useDirection();
 
   const renderPostBottom = (post: Post) => {
     return (
@@ -78,9 +50,7 @@ const Review = () => {
           ))}
         </ReviewWrapper>
 
-        <Float direction={scrollDir}>
-          <Button iconName="pen" text="후기 또는 질문 쓰기" />
-        </Float>
+        <Float scrollDir={scrollDir} distance="1.5rem" />
       </>
     </Layout>
   );
@@ -96,23 +66,6 @@ const ReviewWrapper = styled.div`
     font-weight: 700;
     margin-bottom: 0.5rem;
   }
-`;
-
-interface FloatProps {
-  direction: DIRECTION | undefined;
-}
-
-const Float = styled.div<FloatProps>`
-  display: flex;
-  justify-content: center;
-  position: fixed;
-  bottom: 1.5rem;
-  left: 0;
-  right: 0;
-  z-index: 1000;
-  transition: all 0.3s linear;
-  transform: ${(props) =>
-    props.direction === DIRECTION.Up ? "" : "translateY(10rem)"};
 `;
 
 const PostBottom = styled.div`
